// src/lib/generated/referenceProfiles.ts
export const REFERENCE_CORPUS_METADATA = {
  "corpus": "OpenScore Lieder Corpus",
  "release": "v3.0.0",
  "commit": "f44e962a1ff85ddcc2f6f42bed3b0c529555ee15",
  "license": "CC0-1.0",
  "licenseSha256": "a2010f343487d3f7618affe54f789f5487602331c0a8d03f49e9a7c547cf0499",
  "workCount": 14,
  "manifestSha256": "bf21a3e767c5e97a014684632a2c9aabb1f26691ea886eeb4a03bd07e90b8acf",
  "priorsSha256": "c028699ed6894ed50c7dc76401a94d2a37f40f0b1b92e3a69045478ea311dc2d"
}
// ...
export const REFERENCE_NGRAM_HASHES = [
  "0616e9fc",
  "3d04c8f5",
  "463621d5",
  "a26f0395",
  "b6dc6483",
  "db7a6cd7",
  "ef8413e2",
  "0a049857",
  "36196aef",
  "36a6316c",
  "5512ff12",
  "5cb127b5",
  "608bbcaf",
  "637e8857",
  "6be7d046",
  "74248bef",
  "81e38763",
  "891f04d4",
  "910f7ed4",
  "988e4fc9",
  "9c6efa23",
  "9d6ed40c",
  "a6a70110",
  "afd9dd8c",
  "bbd25858",
  "bf1dc235",
  "cda651b4",
  "dadaafd2",
  "dff4c163",
  "e9840a70",
  "007845f5",
  "008f54c3",
  "012c6611",
  "028948a7",
  "0604920b",
  "08049531",
  "08481401",
  "10a997f0",
  "10e162d5",
  "13707916",
  "15707c3c",
  "16a9a162",
  "1c9836ef",
  "27ae2f52",
  "2accf0ac",
  "2d52d408",
  "30cd98fa",
  "34146223",
  "341967c9",
  "383b70fc",
  "38a48ca6",
  "3936c208",
  "3b6a0738",
  "3be9888d",
  "3ebec5c3",
  "3fc578af",
  "40361863",
  "44b5f4f7",
  "453fac08",
  "47d7f320",
  "49732630",
  "4bc5aeb8",
  "555d967c",
  "5575dfe7",
  "58bf4ede",
  "5aa382c3",
  "5b375565",
  "5bbe77e4",
  "6009d37a",
  "612ed228",
  "633e1568",
  "6850355b",
  "6a951c12",
  "6af9dc63",
  "708a6c1c",
  "722488c9",
  "7327506c",
  "780ec711",
  "7a0eca37",
  "7ac484e2",
  "7ed4f76f",
  "84fb4225",
  "86152763",
  "86e16224",
  "8de39a47",
  "90030a7d",
  "99a7ba7f",
  "9a0c97aa",
  "9a96be2c",
  "9b9489ea",
  "9f86d486",
  "a4e7c442",
  "a8cb4104",
  "a9a622d6",
  "a9dcd26b",
  "aca70a82",
  "ad53e854",
  "b26a7a43",
  "b2c201b2",
  "b33a7c4c",
  "b84cefd6",
  "bbefe360",
  "bf76f728",
  "bffa3628",
  "c0a6cd74",
  "c2ea7ea0",
  "c577009a",
  "c7eff644",
  "cba66337",
  "d0ca5f34",
  "d23d0bc9",
  "d380887c",
  "d4a319a2",
  "d5410816",
  "d61d0303",
  "d97a69b1",
  "dda6b378",
  "df58b9e4",
  "df77edfc",
  "dfc1d7a3",
  "e0ebeefc",
  "e781d001",
  "e963af2a",
  "ebc01951",
  "ed4a48a8",
  "efee7ebe",
  "f5dffd6c",
  "fc22ebb4",
  "fc9b2641",
  "fda9f1c8",
  "feaf60af"
]